+++ crates/zizmor/src/lsp.rs
@@ -190,8 +190,12 @@ impl Backend {
         let mut registry = FindingRegistry::new(&input_registry, None, None, Persona::Regular);
 
         for (input_key, input) in input_registry.iter_inputs() {
-            for (_, audit) in self.audit_registry.iter_audits() {
-                registry.extend(audit.audit(input, input_registry.get_config(input_key.group()))?);
+            for (ident, audit) in self.audit_registry.iter_audits() {
+                registry.extend(audit.audit(
+                    ident,
+                    input,
+                    input_registry.get_config(input_key.group()),
+                )?);
             }
         }